requirements.txt
amqp==2.5.1
anyjson==0.3.3
appdirs==1.4.4
astroid==2.3.0
attrdict==2.0.1
awscli==1.16.219
base58==1.0.3
beautifulsoup4==4.9.3
billiard==3.5.0.5
boto3==1.9.209
botocore==1.12.209
celery==4.1.1
certifi==2019.6.16
chardet==3.0.4
click==7.1.2
colorama==0.3.9
configparser==5.0.1
coreapi==2.3.3
coreschema==0.0.4
crayons==0.4.0
cytoolz==0.10.0
distlib==0.3.0
Django==2.2.4
django-celery==3.3.1
django-filter==2.2.0
django-rest-swagger==2.2.0
django-shell-plus==1.1.7
djangorestframework==3.10.2
docutils==0.14
eth-abi==2.0.0
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.6.4
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.1.0
Flask-PyMongo==2.3.0
grpcio==1.24.3
grpcio-tools==1.24.3
hexbytes==0.2.0
idna==2.8
importlib-metadata==0.19
ipfshttpclient==0.4.12
iroha==0.0.5.4
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonschema==2.6.0
kombu==4.6.4
lazy-object-proxy==1.4.2
lru-dict==1.1.6
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
multiaddr==0.0.8
netaddr==0.7.19
numpy==1.17.2
openapi-codec==1.3.2
parsimonious==0.8.1
protobuf==3.9.1
pyasn1==0.4.6
pycryptodome==3.8.2
PyJWT==2.0.1
pylint==2.4.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
pymongo==3.11.3
pySqsListener==0.8.7
python-dateutil==2.8.0
python-dotenv==0.15.0
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
rlp==1.1.0
rsa==3.4.2
s3transfer==0.2.1
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
soupsieve==2.2
sqlparse==0.3.0
toolz==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
varint==1.0.2
vine==1.3.0
virtualenv==20.0.20
web3==5.0.1
webdriver-manager==3.3.0
websockets==7.0
Werkzeug==1.0.1
wrapt==1.11.2
zipp==0.5.2
